Revert "removed CustomNftCollection.asmdef and moved the service into BackpackEditorHUDV2.asmdef"
This reverts commit ee9e5bbd7a8d1c3314603eb4de01cd63a08f3e8e.
