+++ unity-renderer/Assets/DCLServices/CustomNftCollection.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 91f5002112ab4f389507364c9e8bd920
+timeCreated: 1692900786
