+++ unity-renderer/Assets/DCLServices/CustomNftCollection/CustomNftCollection.asmdef.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d6fe1b5729e3405d98cf8ff2035d9542
+timeCreated: 1692900794
