+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/BackpackEditorHUDV2.asmdef
@@ -70,8 +70,7 @@
         "GUID:a4bf61c057a098f4ca05b539a6d8c0fe",
         "GUID:5ec801bb410ed4adf91f82aa97aaf380",
         "GUID:4ad87ae8afda2df4d897ad1a3527cd1d",
-        "GUID:d6fe1b5729e3405d98cf8ff2035d9542",
-        "GUID:bd51ff2b035ce4ab4be7c38c43c88889"
+        "GUID:d6fe1b5729e3405d98cf8ff2035d9542"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
