+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Environment/Factories/ServiceLocatorFactory/ServiceLocatorFactory.asmdef
@@ -88,8 +88,7 @@
         "GUID:4da9df9a132541f2a0ce3fc6f0578851",
         "GUID:a385efcc15091954c9f00857801f60b3",
         "GUID:d6fe1b5729e3405d98cf8ff2035d9542",
-        "GUID:a4bf61c057a098f4ca05b539a6d8c0fe",
-        "GUID:0c692b52b8dc4a8e9802ddcb8e10f231"
+        "GUID:a4bf61c057a098f4ca05b539a6d8c0fe"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
