+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Environment/Factories/ServiceLocatorFactory/ServiceLocatorFactory.cs
@@ -1,5 +1,4 @@
 using AvatarSystem;
-using DCL.Backpack;
 using DCL.Chat;
 using DCL.Chat.Channels;
 using DCL.Controllers;
@@ -14,6 +13,7 @@ using DCL.Social.Friends;
 using DCL.World.PortableExperiences;
 using DCLServices.CameraReelService;
 using DCLServices.CopyPaste.Analytics;
+using DCLServices.CustomNftCollection;
 using DCLServices.DCLFileBrowser;
 using DCLServices.DCLFileBrowser.DCLFileBrowserFactory;
 using DCLServices.EmotesCatalog;
