+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Environment/TestEnvironment/TestEnvironment.asmdef
@@ -47,8 +47,7 @@
         "GUID:5db7e9d450464b96bfddf316690db751",
         "GUID:70263d10ecfb462ba0f6762dda35634b",
         "GUID:4da9df9a132541f2a0ce3fc6f0578851",
-        "GUID:d6fe1b5729e3405d98cf8ff2035d9542",
-        "GUID:0c692b52b8dc4a8e9802ddcb8e10f231"
+        "GUID:d6fe1b5729e3405d98cf8ff2035d9542"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
